--- apps/backend-e2e/project.json
@@ -1,22 +1,0 @@
-{
-  "name": "backend-e2e",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "implicitDependencies": ["backend"],
-  "targets": {
-    "e2e": {
-      "executor": "@nx/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/{e2eProjectRoot}"],
-      "options": {
-        "jestConfig": "apps/backend-e2e/jest.config.ts",
-        "passWithNoTests": true
-      }
-    },
-    "lint": {
-      "executor": "@nx/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["apps/backend-e2e/**/*.{js,ts}"]
-      }
-    }
-  }
-}
