--- apps/frontend-e2e/cypress.config.ts
@@ -1,6 +1,0 @@
-import { defineConfig } from 'cypress';
-import { nxE2EPreset } from '@nx/cypress/plugins/cypress-preset';
-
-export default defineConfig({
-  e2e: nxE2EPreset(__dirname),
-});
