+++ libs/database/README.md
@@ -1,3 +1,7 @@
 # database
 
 This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test database` to execute the unit tests via [Jest](https://jestjs.io).
