--- libs/database/config/config.ts
@@ -1,5 +1,0 @@
-import { registerAs } from '@nestjs/config';
-
-export default registerAs('mongoDb', () => ({
-  databaseUrl: process.env.DATABASE_URL,
-}));
