--- libs/database/database/database.schema.ts
@@ -1,18 +1,0 @@
-import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
-import { HydratedDocument } from 'mongoose';
-
-export type CatDocument = HydratedDocument<Cat>;
-
-@Schema()
-export class Cat {
-  @Prop()
-  name: string;
-
-  @Prop()
-  age: number;
-
-  @Prop()
-  breed: string;
-}
-
-export const CatSchema = SchemaFactory.createForClass(Cat);
