--- libs/database/index.ts
@@ -1,2 +1,0 @@
-export * from './config';
-export * from './database';
