+++ libs/database/src/database.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { PrismaModule } from './prisma/prisma.module';
+
+@Module({
+  providers: [PrismaModule],
+  exports: [PrismaModule],
+  imports: [PrismaModule],
+})
+export class DatabaseModule {}
