+++ libs/database/src/index.ts
@@ -1,2 +1,1 @@
 export * from './database.module';
-export * from './database.schema';
