+++ libs/database/src/prisma/prisma.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { PrismaService } from './prisma.service';
+
+@Module({
+  imports: [],
+  exports: [PrismaService],
+  providers: [PrismaService],
+})
+export class PrismaModule {}
