+++ nest-cli.json
@@ -1,0 +1,25 @@
+{
+  "projects": {
+    "prisma": {
+      "type": "library",
+      "root": "libs/database/src/prisma",
+      "entryFile": "index",
+      "sourceRoot": "libs/database/src/prisma/src",
+      "compilerOptions": {
+        "tsConfigPath": "libs/database/src/prisma/tsconfig.lib.json"
+      }
+    },
+    "database": {
+      "type": "library",
+      "root": "libs/database",
+      "entryFile": "index",
+      "sourceRoot": "libs/database/src",
+      "compilerOptions": {
+        "tsConfigPath": "libs/database/tsconfig.lib.json"
+      }
+    }
+  },
+  "compilerOptions": {
+    "webpack": true
+  }
+}
